Improving documentation for Vertex Normals
Looking at: "Table of Indices, Positions and Normals for Flat Shaded Box" there is a table.
In the table there are 24 rows. But when I run the playground example and add a few console logs it will produce a table with 36 rows.

Example - https://www.babylonjs-playground.com/#1H7L5C#64

Changing the table in the documentation to reflect the values produced by the example that is included with the documentation.

PR is based on discussion with Deltakosh at https://forum.babylonjs.com/t/vertex-normals-babylon-js-documentation-has-a-table-with-24-rows-and-they-probably-should-be-36/15603