Create a top-level Contributing page
IMHO, this is critical information that will be of interest to anyone who might consider becoming an active contributor, especially new users who haven't contributed before. Hiding it deep down in the hierarchy of documentation pages seems unwise. Removing barriers to entry like this should increase the amount and quality of contributions that Babylon can attract.