Increase the contrast of the Table of Contents component
This makes the TOC stand out visually and the text easier to read. The padding is derived from the padding-bottom that was already in place, though I couldn't find a way to access that value directly (which would be preferable in case of future changes).

The heading uses a 13px margin so that's a bit inconsistent, but I'm not sure if it makes much of a difference visually and so I've left it as-is.