Call `initializeBabylonApp(...)` in Home.tsx
The Home.tsx example is missing the call to `initializeBabylonApp(babylonOptions)` after the options object is initialized. This makes the Ionic app show nothing but a white screen when the `ionic serve` command is run.

This change fixes the issue by adding the call to `initializeBabylonApp(babylonOptions)` immediately after the options object is initialized.