Update h tags in text and h styles in scss
Updated headers to have the correct h tag level for ToC indention. Also updated scss for h tags to have more contrast  and visual balance between the levels. Increased the size of h5 and h6 to not be smaller than p tag text as well as giving some styling emphasis.