Updates from Alexis's second review + matrix compose
These are updates from Alexis's second review as well as a full spell check. Also completed the Align entry as well as updated the AlignNode binary to push up again as it did not appear to commit with the original PR. Also added the Matrix Compose docs as it's a new node that has come in since the original PR and it seemed logical to add it now.