Update node anchor links with new tag
Update to the tableOfContentItem component with direction from Raanan to fix a missing line in the H3 tag code.

Update the nodes doc to exchange the image and h3 tag with the new H3 tag to allow anchor linking to the image rather than to the title of the node which scrolls the image of the node off the top of the screen.