Updating PG search to handle PG 2 manifest (#1402)
This PR updates the Playground search page to handle the new manifest
that the Playground V2 uses (from this PR:
https://github.com/BabylonJS/Babylon.js/pull/17175).

The schema of the manifest of the PG V2 is defined as:
```
export type V2Manifest = {
    v: 2;
    language: "JS" | "TS";
    entry: string;
    imports: Record<string, string>;
    files: Record<string, string>;
    cdnBase?: string;
};
```

For search, we care about the _files_ property, which contains pairs
with the name of the file and the code of the file. Most of the parsing
changes are related to dealing with the fact that now there are multiple
files to search in a PG and to surface the name of the file where the
code is found (if we are searching by Code).

All search behavior is identical for V1 and V2, except for the following
changes:
- If the playground is a V2 playground, when searching for code with an
exact match the name of the file will be printed with the results:
<img width="1869" height="1069" alt="image"
src="https://github.com/user-attachments/assets/72d96489-08e8-41ae-a065-08c3dee4abbf"
/>
- Ellipsis in the code are removed for long lines, this is done to
simplify the code that handles finding the correct line and the current
page handles long lines correctly by wrapping the lines.